integrate transaction-missing into Support screen
