+++ src/App.tsx
@@ -31,7 +31,6 @@ const SellInfoScreen = lazy(() => import('./screens/sell-info.screen'));
 const SupportIssueScreen = lazy(() => import('./screens/support-issue.screen'));
 const SupportScreen = lazy(() => import('./screens/support.screen'));
 const TfaScreen = lazy(() => import('./screens/tfa.screen'));
-const TransactionMissingScreen = lazy(() => import('./screens/transaction-missing.screen'));
 const TransactionScreen = lazy(() => import('./screens/transaction.screen'));
 
 setupLanguages();
@@ -149,18 +148,6 @@ export const Routes = [
     path: '/support/issue',
     element: withSuspense(<SupportIssueScreen />),
   },
-  // {
-  //   path: '/support/issue/tx',
-  //   element: withSuspense(<TransactionScreen />),
-  // },
-  {
-    path: '/support/issue/tx-missing',
-    element: withSuspense(<TransactionMissingScreen />),
-  },
-  // {
-  //   path: '/support/issue/tx/:id',
-  //   element: withSuspense(<SupportIssueScreen />),
-  // },
   {
     path: '/bank-accounts',
     element: withSuspense(<BankAccountsScreen />),
